Roll back evaluation code to a running state.
